requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.10.0
httpx>=0.27.0
google-genai>=1.0.0
requests>=2.32.0
gunicorn>=23.0.0
